fish: replace e function with abbreviation
There were several issues with the function including opening spaced
files etc. This is solved by using an abbreviation. The only downside is
that the abbreviation needs to be updated after you switch the editor
but that will probably not happen very often.

That would be solved by quickly running `fish_setup`.