emacs: improve handling of whitespace trim
I don't want to trim whitespace when I am in the middle of writing
something on a line but happens to toggle window or open the ibus
switcher. Since I have auto-save on window inactive this white-space
trimming was too obtrusive. ws-butler is much more forgiving and only
removes the parts that you have edited and not the lines you are
currently editing.