xmonad: improve pip window
Fix annoying toggle on different desktop breaking it. Minimize was not
sane to use in this way since it is per workspace or something like
that. Instead utilize the NSP (named scratchpad) workspace and toggle it
in the same manner.

Fix the clone to all desktops which works better now with the new hooks.

Also mark boring as much as possible, since this also seems to be per workspace.